RFT: move various permutations to utils
