+++ permute/core.py
@@ -15,62 +15,6 @@ from scipy.stats import (binom, ttest_ind)
 from .utils import get_prng
 
 
-def permute_within_groups(x, group, seed=None):
-    """
-    Permutation of condition within each group.
-
-    Parameters
-    ----------
-    x : array-like
-        A 1-d array indicating treatment.
-    group : array-like
-        A 1-d array indicating group membership
-    seed : RandomState instance or {None, int, RandomState instance}
-        If None, the pseudorandom number generator is the RandomState
-        instance used by `np.random`;
-        If int, seed is the seed used by the random number generator;
-        If RandomState instance, seed is the pseudorandom number generator
-
-    Returns
-    -------
-    permuted : array-like
-        The within group permutation of x.
-    """
-    prng = get_prng(seed)
-    permuted = x.copy()
-
-    # (avoid additional flops) -- maybe memoize
-    for g in np.unique(group):
-        gg = group == g
-        permuted[gg] = prng.permutation(permuted[gg])
-    return permuted
-
-
-def permute_rows(m, seed=None):
-    """
-    Permute the rows of a matrix in-place
-
-    Parameters
-    ----------
-    m : array-like
-        A 2-d array
-    seed : RandomState instance or {None, int, RandomState instance}
-        If None, the pseudorandom number generator is the RandomState
-        instance used by `np.random`;
-        If int, seed is the seed used by the random number generator;
-        If RandomState instance, seed is the pseudorandom number generator
-
-    Returns
-    -------
-    None
-        Original matrix is permute in-place, nothing returned
-    """
-    prng = get_prng(seed)
-
-    for row in m:
-        prng.shuffle(row)
-
-
 def corr(x, y, reps=10**4, seed=None):
     """
     Simulate permutation p-value for Spearman correlation coefficient
