+++ permute/tests/test_core.py
@@ -8,27 +8,7 @@ from numpy.random import RandomState
 
 from ..core import (binom_conf_interval,
                     corr,
-                    permute_within_groups,
-                    permute_rows,
                     two_sample)
-
-
-def test_permute_within_group():
-    x = np.repeat([1, 2, 3]*3, 3)
-    group = np.repeat([1, 2, 3], 9)
-    #response = np.zeros_like(group)
-    #response[[0,  1,  3,  9, 10, 11, 18, 19, 20]] = 1
-
-    prng1 = RandomState(42)
-    prng2 = RandomState(42)
-    res1 = permute_within_groups(x, group, prng1)
-    res2 = permute_within_groups(x, group, prng2)
-    np.testing.assert_equal(res1, res2)
-
-    res3 = permute_within_groups(x, group)
-    np.testing.assert_equal(res3.max(), 3)
-    res3.sort()
-    np.testing.assert_equal(group, res3)
 
 
 def test_corr():
@@ -67,20 +47,6 @@ def test_binom_conf_interval():
     res3 = binom_conf_interval(10, 5, cl=0.95, alternative="lower")
     expected3 = (0.22244110100812578, 1.0)
     np.testing.assert_equal(res3, expected3)
-
-
-def test_permute_rows():
-    prng = RandomState(42)
-
-    x = prng.randint(10, size=20).reshape(2, 10)
-    permute_rows(x, prng)
-    expected = np.array([[2, 7, 7, 6, 4, 9, 3, 4, 6, 6],
-                         [7, 4, 5, 5, 3, 7, 1, 2, 7, 1]])
-    np.testing.assert_array_equal(x, expected)
-
-    permute_rows(x)
-    np.testing.assert_equal(x.max(), 9)
-    np.testing.assert_equal(x.min(), 1)
 
 
 @attr('slow')
